+++ core/src/main/kotlin/org/pastalab/fray/core/concurrency/operations/ObjectWakeBlocked.kt
@@ -6,10 +6,10 @@ import org.pastalab.fray.rmi.ResourceType
 
 class ObjectWakeBlocked(val objectContext: ObjectNotifyContext, val noTimeout: Boolean) :
     BlockedOperation(
-        false,
         ResourceInfo(
             System.identityHashCode(objectContext.lockContext.lockReference.get()),
-            ResourceType.CONDITION)) {
+            ResourceType.CONDITION),
+        BLOCKED_OPERATION_NOT_TIMED) {
   override fun unblockThread(tid: Long, type: InterruptionType): Any? {
     if (objectContext.lockContext.canLock(tid)) {
       return objectContext.getSyncObject()