+++ core/src/main/kotlin/org/pastalab/fray/core/concurrency/operations/ReactiveBlockingBlocked.kt
@@ -1,15 +1,13 @@
-package org.pastalab.fray.core.concurrency.context
+package org.pastalab.fray.core.concurrency.operations
 
 import org.pastalab.fray.core.ThreadContext
-import org.pastalab.fray.core.concurrency.operations.BlockedOperation
-import org.pastalab.fray.core.concurrency.operations.InterruptionType
 import org.pastalab.fray.rmi.ResourceInfo
+import org.pastalab.fray.rmi.ResourceType
 
 class ReactiveBlockingBlocked(val threadContext: ThreadContext) :
     BlockedOperation(
-        false,
-        ResourceInfo(
-            System.identityHashCode(threadContext), org.pastalab.fray.rmi.ResourceType.NETWORK)) {
+        ResourceInfo(System.identityHashCode(threadContext), ResourceType.NETWORK),
+        BLOCKED_OPERATION_NOT_TIMED) {
   override fun unblockThread(tid: Long, type: InterruptionType): Any? {
     threadContext.interruptSignaled = true
     threadContext.thread.interrupt()