+++ core/src/main/kotlin/org/pastalab/fray/core/concurrency/operations/SocketChannelReadBlocked.kt
@@ -5,7 +5,7 @@ import org.pastalab.fray.rmi.ResourceInfo
 import org.pastalab.fray.rmi.ResourceType
 
 class SocketChannelReadBlocked(val socketChannelContext: SocketChannelContext) :
-    BlockedOperation(false, ResourceInfo(0, ResourceType.NETWORK)) {
+    BlockedOperation(ResourceInfo(0, ResourceType.NETWORK), BLOCKED_OPERATION_NOT_TIMED) {
   override fun unblockThread(tid: Long, type: InterruptionType): Any? {
     return null
   }